requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
APScheduler==3.10.4
SQLAlchemy==2.0.23
pandas==2.1.3
plotly==5.18.0
streamlit==1.28.1
gensim==4.3.2
underthesea==1.3.3
python-dotenv==1.0.0
torch==2.1.2
transformers==4.36.2
scikit-learn==1.3.2
numpy==1.24.3
tqdm==4.66.1
bertopic>=0.16.0
accelerate>=0.26.0